fix pixi format

+++ problems/p21/op/embedding.mojo
@@ -114,11 +114,14 @@ struct EmbeddingCustomOp:
         embed_dim: Int,
     ](
         output: OutputTensor[
-            dtype = DType.float32, rank=3, static_spec=_],  # [batch_size, seq_len, embed_dim]
+            dtype=DType.float32, rank=3, static_spec=_
+        ],  # [batch_size, seq_len, embed_dim]
         indices: InputTensor[
-            dtype = DType.int32, rank=2, static_spec=_],  # [batch_size, seq_len]
+            dtype=DType.int32, rank=2, static_spec=_
+        ],  # [batch_size, seq_len]
         weights: InputTensor[
-            dtype = output.dtype, rank=2, static_spec=_],  # [vocab_size, embed_dim]
+            dtype=output.dtype, rank=2, static_spec=_
+        ],  # [vocab_size, embed_dim]
         ctx: DeviceContextPtr,
     ) raises:
         var output_tensor = output.to_layout_tensor()
@@ -193,11 +196,14 @@ struct Embedding2DCustomOp:
         embed_dim: Int,
     ](
         output: OutputTensor[
-            dtype = DType.float32, rank=3, static_spec=_],  # [batch_size, seq_len, embed_dim]
+            dtype=DType.float32, rank=3, static_spec=_
+        ],  # [batch_size, seq_len, embed_dim]
         indices: InputTensor[
-            dtype = DType.int32, rank=2, static_spec=_],  # [batch_size, seq_len]
+            dtype=DType.int32, rank=2, static_spec=_
+        ],  # [batch_size, seq_len]
         weights: InputTensor[
-            dtype = output.dtype, rank=2, static_spec=_],  # [vocab_size, embed_dim]
+            dtype=output.dtype, rank=2, static_spec=_
+        ],  # [vocab_size, embed_dim]
         ctx: DeviceContextPtr,
     ) raises:
         var output_tensor = output.to_layout_tensor()
